#1134 Prepare Email models

+++ database/migrations/2023_03_08_010854_create_mailshot_stats_table.php
@@ -16,7 +16,7 @@ return new class () extends Migration {
     {
         Schema::create('mailshot_stats', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedSmallInteger('mailshot_id')->nullable();
+            $table->unsignedInteger('mailshot_id')->nullable();
             $table->foreign('mailshot_id')->references('id')->on('mailshots');
             $table = $this->dispatchedEmailStats($table);
